Move testcases in test_templated_script_task.py in idmtools_models for comps test to idmtools_platform_comps/tests folder
