+++ idmtools_models/tests/test_templated_script_task.py
@@ -3,7 +3,6 @@ import os
 import sys
 from unittest import TestCase
 import pytest
-from idmtools.assets import Asset
 from idmtools.core.platform_factory import Platform
 from idmtools.entities import CommandLine
 from idmtools.entities.command_task import CommandTask
@@ -180,94 +179,6 @@ echo Hello
                 self.assertEqual(1, experiment.assets.count)
                 self.assertIn("wrapper.sh", str(task.command))
 
-    @pytest.mark.comps
-    @pytest.mark.timeout(60)
-    def test_wrapper_script_execute_comps(self):
-        """
-        This tests The ScriptWrapperScriptTask as well as the TemplatedScriptTask
-
-        In addition, it tests reload
-        Returns:
-
-        """
-        cmd = "python3.6 --version"
-        task = CommandTask(cmd)
-        task.common_assets.add_asset(
-            Asset(relative_path=os.path.join("site-packages", "test-package"), filename="__init__.py", content="a=\'123\'"))
-        pl_slurm = Platform("SLURM")
-        wrapper_task: TemplatedScriptTask = get_script_wrapper_unix_task(task, template_content=LINUX_PYTHON_PATH_WRAPPER)
-        wrapper_task.script_binary = "/bin/bash"
-        experiment = Experiment.from_task(wrapper_task, name=self.case_name)
-        experiment.run(wait_until_done=True)
-        self.assertTrue(experiment.succeeded)
-
-        for sim in experiment.simulations:
-            assets = pl_slurm._simulations.all_files(sim)
-            for asset in assets:
-                content = asset.content.decode('utf-8').replace("\\\\", "\\")
-                if asset.filename in ["stdout.txt"]:
-                    # check for echo
-                    self.assertIn('Running', content)
-                    # don't check full version in case comps updates system
-                    self.assertIn('Python 3.6', content)
-
-    @pytest.mark.comps
-    def test_wrapper_script_python_execute_site_packages_comps(self):
-        """
-        This tests The ScriptWrapperScriptTask as well as the TemplatedScriptTask
-
-        In addition, it tests reload
-        Returns:
-
-        """
-        cmd = "python3.6 -c \"import test_package as tp;print(tp.a)\""
-        task = CommandTask(cmd)
-        task.common_assets.add_asset(Asset(relative_path=os.path.join("site-packages", "test_package"), filename="__init__.py", content="a=\'123\'"))
-        pl_slurm = Platform("SLURM")
-        wrapper_task: TemplatedScriptTask = get_script_wrapper_unix_task(task, template_content=LINUX_PYTHON_PATH_WRAPPER)
-        wrapper_task.script_binary = "/bin/bash"
-        experiment = Experiment.from_task(wrapper_task, name=self.case_name)
-        experiment.run(wait_until_done=True)
-        self.assertTrue(experiment.succeeded)
-
-        for sim in experiment.simulations:
-            assets = pl_slurm._simulations.all_files(sim)
-            for asset in assets:
-                content = asset.content.decode('utf-8').replace("\\\\", "\\")
-                if asset.filename in ["stdout.txt"]:
-                    # check for echo
-                    self.assertIn('Running', content)
-                    # don't check full version in case comps updates system
-                    self.assertIn('123', content)
-
-    @pytest.mark.comps
-    def test_wrapper_script_python_execute_assets_comps(self):
-        """
-        This tests The ScriptWrapperScriptTask as well as the TemplatedScriptTask
-
-        In addition, it tests reload
-        Returns:
-
-        """
-        cmd = "python3.6 -c \"import test_package as tp;print(tp.a)\""
-        task = CommandTask(cmd)
-        task.common_assets.add_asset(Asset(filename="test_package.py", content="a=\'123\'"))
-        pl_slurm = Platform("SLURM")
-        wrapper_task: TemplatedScriptTask = get_script_wrapper_unix_task(task, template_content=LINUX_PYTHON_PATH_WRAPPER)
-        wrapper_task.script_binary = "/bin/bash"
-        experiment = Experiment.from_task(wrapper_task, name=self.case_name)
-        experiment.run(wait_until_done=True)
-        self.assertTrue(experiment.succeeded)
-
-        for sim in experiment.simulations:
-            assets = pl_slurm._simulations.all_files(sim)
-            for asset in assets:
-                content = asset.content.decode('utf-8').replace("\\\\", "\\")
-                if asset.filename in ["stdout.txt"]:
-                    # check for echo
-                    self.assertIn('Running', content)
-                    # don't check full version in case comps updates system
-                    self.assertIn('123', content)
